Update COCOMetric to use MMEval
