+++ mmpose/evaluation/metrics/__init__.py
@@ -1,12 +1,12 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .coco_metric import AP10KCocoMetric, CocoMetric
+from .coco_metric import COCOMetric
+from .coco_metric_old import CocoMetric
 from .coco_wholebody_metric import CocoWholeBodyMetric
 from .keypoint_2d_metrics import (AUC, EPE, NME, JhmdbPCKAccuracy,
                                   MpiiPCKAccuracy, PCKAccuracy)
 from .posetrack18_metric import PoseTrack18Metric
 
 __all__ = [
-    'CocoMetric', 'AP10KCocoMetric', 'PCKAccuracy', 'MpiiPCKAccuracy',
-    'JhmdbPCKAccuracy', 'AUC', 'EPE', 'NME', 'PoseTrack18Metric',
-    'CocoWholeBodyMetric'
+    'COCOMetric', 'PCKAccuracy', 'MpiiPCKAccuracy', 'JhmdbPCKAccuracy', 'AUC',
+    'EPE', 'NME', 'PoseTrack18Metric', 'CocoWholeBodyMetric', 'CocoMetric'
 ]
